+++ mocks/include/mocks/EmptyApplication.hpp
@@ -75,6 +75,12 @@ public:
         return nullptr;
     }
 
+    Logging *getChatLogger() override
+    {
+        assert(!"getChatLogger was called without being initialized");
+        return nullptr;
+    }
+
     ChatterinoBadges *getChatterinoBadges() override
     {
         return nullptr;